requirements.txt
asyncio
python-telegram-bot==13.7
requests
flask